% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_prob.R
\name{joint_prob}
\alias{joint_prob}
\title{Calculate Joint Probability}
\usage{
joint_prob(xe, xs, n, p)
}
\arguments{
\item{xe}{Integer. The observed value of the random variable \( X_e \)}

\item{xs}{Integer. The observed value of the random variable \( X_s \).}

\item{n}{Integer. The total number of trials or sample size.}

\item{p}{Numeric vector. A vector of length 4 representing the probabilities associated with each outcome in the multinomial distribution. The elements of \( p \) should sum to 1.}
}
\value{
Numeric. The joint probability \( P(X_e = x_e, X_s = x_s) \).
}
\description{
This function computes the joint probability of observing specific values \( x_e \) and \( x_s \) for the random variables \( X_e \) and \( X_s \), respectively, given the total number of trials \( n \) and a probability vector \( p \). The probability vector \( p \) should be of length 4, representing the probabilities of all possible outcomes in a multinomial distribution.
}
\examples{
joint_prob(2,2,5,c(0.1,0.1,0.2,0.6))



}
