% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob}
\alias{prob}
\title{Calculate Multinomial Distribution Parameters}
\usage{
prob(pr, pt, fi)
}
\arguments{
\item{pr}{Numeric. Probability of efficacy (\code{pe}), where \code{0 <= pe <= 1}.}

\item{pt}{Numeric. Probability of safety (\code{ps}), where \code{0 <= ps <= 1}.}

\item{fi}{Numeric. Odds ratio (\code{phi}) representing the relationship between efficacy and safety outcomes, defined as \code{phi = (p1 * p4) / (p2 * p3)}.}
}
\value{
A numeric vector of length 4 representing the multinomial distribution parameters \code{[p1, p2, p3, p4]}.
}
\description{
Computes the parameters of a multinomial distribution represented by a 4-element vector \code{p = [p1, p2, p3, p4]}.
The function takes the probability of efficacy (\code{pe}), the probability of safety (\code{ps}), and the odds ratio (\code{phi})
to determine the joint probabilities under the constraints:

\itemize{
\item \code{p1 + p2 = pe}
\item \code{p1 + p3 = ps}
\item \code{p1 + p2 + p3 + p4 = 1}
}
}
\examples{
prob(0.7,0.5,1.5)

}
